+++ mod-source-record-manager-server/src/main/java/org/folio/services/JobExecutionServiceImpl.java
@@ -5,6 +5,7 @@ import io.vertx.core.Promise;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.handler.HttpException;
+
 import org.apache.commons.io.FilenameUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -25,6 +26,7 @@ import org.folio.rest.jaxrs.model.InitJobExecutionsRqDto;
 import org.folio.rest.jaxrs.model.InitJobExecutionsRsDto;
 import org.folio.rest.jaxrs.model.JobExecution;
 import org.folio.rest.jaxrs.model.JobExecutionDtoCollection;
+import org.folio.rest.jaxrs.model.JobExecutionUserInfoCollection;
 import org.folio.rest.jaxrs.model.JobProfile;
 import org.folio.rest.jaxrs.model.JobProfileInfo;
 import org.folio.rest.jaxrs.model.ProfileSnapshotWrapper;
@@ -33,11 +35,14 @@ import org.folio.rest.jaxrs.model.RunBy;
 import org.folio.rest.jaxrs.model.Snapshot;
 import org.folio.rest.jaxrs.model.StatusDto;
 import org.folio.rest.jaxrs.model.UserInfo;
+import org.folio.rest.jaxrs.model.JobProfileInfoCollection;
+import org.folio.services.exceptions.JobDuplicateUpdateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.NotFoundException;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -201,6 +206,11 @@ public class JobExecutionServiceImpl implements JobExecutionService {
         }));
   }
 
+  @Override
+  public Future<JobProfileInfoCollection> getRelatedJobProfiles(int offset, int limit, String tenantId) {
+    return jobExecutionDao.getRelatedJobProfiles(offset, limit, tenantId);
+  }
+
   private Future<ProfileSnapshotWrapper> createJobProfileSnapshotWrapper(JobProfileInfo jobProfile, OkapiConnectionParams params) {
     Promise<ProfileSnapshotWrapper> promise = Promise.promise();
     DataImportProfilesClient client = new DataImportProfilesClient(params.getOkapiUrl(), params.getTenantId(), params.getToken());
@@ -242,12 +252,22 @@ public class JobExecutionServiceImpl implements JobExecutionService {
       .map(this::modifyJobExecutionToCompleteWithCancelledStatus)
       .compose(jobExec -> updateJobExecutionWithSnapshotStatus(jobExec, params))
       .compose(jobExec -> deleteRecordsFromSRSIfNecessary(jobExec, params))
-      .map(true);
+      .map(true)
+      .recover(
+        throwable -> throwable instanceof JobDuplicateUpdateException ?
+          Future.succeededFuture(true) :
+          Future.failedFuture(throwable)
+      );
   }
 
   @Override
-  public Future<DeleteJobExecutionsResp>  softDeleteJobExecutionsByIds(List<String> ids, String tenantId) {
+  public Future<DeleteJobExecutionsResp> softDeleteJobExecutionsByIds(List<String> ids, String tenantId) {
     return jobExecutionDao.softDeleteJobExecutionsByIds(ids, tenantId);
+  }
+
+  @Override
+  public Future<JobExecutionUserInfoCollection> getRelatedUsersInfo(int offset, int limit, String tenantId) {
+    return jobExecutionDao.getRelatedUsersInfo(offset, limit, tenantId);
   }
 
   /**
@@ -484,8 +504,8 @@ public class JobExecutionServiceImpl implements JobExecutionService {
     if (jobExecution.getStatus() == JobExecution.Status.ERROR || jobExecution.getStatus() == COMMITTED
     || jobExecution.getStatus() == JobExecution.Status.CANCELLED) {
       String msg = String.format("JobExecution with status '%s' cannot be forcibly completed", jobExecution.getStatus());
-      LOGGER.error(msg);
-      throw new BadRequestException(msg);
+      LOGGER.info(msg);
+      throw new JobDuplicateUpdateException(msg);
     }
     return jobExecution;
   }