+++ mod-source-record-manager-server/src/main/java/org/folio/services/JournalRecordService.java
@@ -1,7 +1,6 @@
 package org.folio.services;
 
 import io.vertx.core.Future;
-import org.folio.rest.jaxrs.model.JobExecutionLogDto;
 import org.folio.rest.jaxrs.model.JobExecutionSummaryDto;
 import org.folio.rest.jaxrs.model.JobLogEntryDtoCollection;
 import org.folio.rest.jaxrs.model.JournalRecordCollection;
@@ -24,15 +23,6 @@ public interface JournalRecordService {
   Future<Boolean> deleteByJobExecutionId(String jobExecutionId, String tenantId);
 
   /**
-   * Returns JobExecutionLogDto with import results for job execution with specified id
-   *
-   * @param jobExecutionId jobExecution id
-   * @param tenantId       tenant id
-   * @return future with JobExecutionLogDto entity
-   */
-  Future<JobExecutionLogDto> getJobExecutionLogDto(String jobExecutionId, String tenantId);
-
-  /**
    * Searches for JournalRecords by jobExecutionId and sorts them using specified sort criteria and direction
    *
    * @param jobExecutionId job execution id
@@ -49,12 +39,13 @@ public interface JournalRecordService {
    * @param jobExecutionId job execution id
    * @param sortBy         sorting criteria
    * @param order          sorting direction
+   * @param errorsOnly     filtering by error field
    * @param limit          limit
    * @param offset         offset
    * @param tenantId       tenantId
    * @return future with JobLogEntryDto collection
    */
-  Future<JobLogEntryDtoCollection> getJobLogEntryDtoCollection(String jobExecutionId, String sortBy, String order, int limit, int offset, String tenantId);
+  Future<JobLogEntryDtoCollection> getJobLogEntryDtoCollection(String jobExecutionId, String sortBy, String order, boolean errorsOnly, int limit, int offset, String tenantId);
 
   /**
    * Searches for RecordProcessingLogDto entity by jobExecutionId and recordId